+++ example/testcontent/models.py
@@ -6,6 +6,7 @@ from elasticsearch_dsl import field
 from djbetty.fields import ImageField
 
 from bulbs.content.models import Content, Tag, ElasticsearchImageField
+from bulbs.liveblog.models import AbstractLiveBlog
 from bulbs.reading_list.mixins import ReadingListMixin
 from bulbs.recirc.mixins import BaseQueryMixin
 from bulbs.videos.mixins import VideoMixin
@@ -104,3 +105,12 @@ class TestVideoContentObj(Content, VideoMixin):
 
     def get_absolute_url(self):
         return '/detail/%s/' % self.pk
+
+
+class TestLiveBlog(Content, AbstractLiveBlog):
+    """Example LiveBlog implementation"""
+
+    class Mapping(Content.Mapping):
+        class Meta:
+            # Exclude until actually needed, to avoid dealing with custom mappings
+            excludes = ('pinned_content', 'recirc_content')